Move stream filter to helpers
